chore (apollo): rename mock-db to mock-db.js
