+++ src/models/destination.model.js
@@ -1,4 +1,4 @@
-import { destinations } from '../../db';
+import { destinations } from '../../mock-db';
 
 export default {
   getAll : () => Object.values(destinations),
